Improve layout for auto reload.
