+++ frontend/.eslintrc
@@ -40,7 +40,7 @@
     //"no-console": 2,
     "no-loops/no-loops": "error",
     "react-refresh/only-export-components": [
-      "warn",
+      "error",
       { "allowConstantExport": true }
     ],
     "react/prop-types": [
