+++ frontend/package.json
@@ -56,8 +56,8 @@
     "@types/jest": "^29.5.12",
     "@types/loadable__component": "^5.13.8",
     "@types/lodash": "^4.14.202",
-    "@types/node": "^20.11.19",
-    "@types/react": "^18.2.57",
+    "@types/node": "^20.11.20",
+    "@types/react": "^18.2.58",
     "@types/react-dom": "^18.2.19",
     "@types/steamid": "^2.0.3",
     "@typescript-eslint/eslint-plugin": "^7.0.2",
@@ -74,7 +74,7 @@
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
-    "i18next": "^23.9.0",
+    "i18next": "^23.10.0",
     "i18next-browser-languagedetector": "^7.2.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
