+++ frontend/src/component/SettingsCheckbox.tsx
@@ -1,0 +1,36 @@
+import { Checkbox, FormControlLabel, FormGroup } from '@mui/material';
+import Tooltip from '@mui/material/Tooltip';
+
+interface SettingsCheckBoxProps {
+    label: string;
+    enabled: boolean;
+    setEnabled: (checked: boolean) => void;
+    tooltip: string;
+}
+
+const SettingsCheckBox = ({
+    label,
+    enabled,
+    setEnabled,
+    tooltip
+}: SettingsCheckBoxProps) => {
+    return (
+        <FormGroup>
+            <Tooltip title={tooltip} placement="top">
+                <FormControlLabel
+                    control={
+                        <Checkbox
+                            checked={enabled}
+                            onChange={(_, checked) => {
+                                setEnabled(checked);
+                            }}
+                        />
+                    }
+                    label={label}
+                />
+            </Tooltip>
+        </FormGroup>
+    );
+};
+
+export default SettingsCheckBox;
