+++ frontend/src/component/menu/NotesMenu.tsx
@@ -31,3 +31,5 @@ export const NotesMenu = ({
         />
     );
 };
+
+export default NotesMenu;
