+++ frontend/src/component/modal/NoteEditorModal.tsx
@@ -11,7 +11,7 @@ import { Trans, useTranslation } from 'react-i18next';
 import { saveUserNote } from '../../api.ts';
 import NiceModal, { muiDialog, useModal } from '@ebay/nice-modal-react';
 import { logError } from '../../util.ts';
-import { CancelButton, ClearButton, SaveButton } from '../Buttons.tsx';
+import { CancelButton, ClearButton, SaveButton } from '../CancelButton.tsx';
 
 interface NoteEditorProps {
     notes: string;
