+++ frontend/src/util.ts
@@ -24,6 +24,10 @@ export const openInNewTab = (url: string) => {
     window.open(url, '_blank');
 };
 
+export interface onClickProps {
+    onClick: () => void;
+}
+
 export const formatExternalLink = (steam_id: string, link: Link): string => {
     const sid = new SteamID(steam_id);
     switch (link.id_format) {
@@ -65,3 +69,45 @@ export const noop = (): void => {};
 export const uniqCI = (values: string[]): string[] => [
     ...new Map(values.map((s) => [s.toLowerCase(), s])).values()
 ];
+
+export const validatorSteamID = (value: string): string => {
+    let err = 'Invalid SteamID';
+    try {
+        const id = new SteamID(value);
+        if (id.isValid()) {
+            err = '';
+        }
+    } catch (_) {
+        /* empty */
+    }
+    return err;
+};
+
+export const makeValidatorLength = (length: number): inputValidator => {
+    return (value: string): string => {
+        if (value.length != length) {
+            return 'Invalid value';
+        }
+        return '';
+    };
+};
+
+export const validatorAddress = (value: string): string => {
+    const pcs = value.split(':');
+    if (pcs.length != 2) {
+        return 'Format must match host:port';
+    }
+    if (pcs[0].toLowerCase() != 'localhost' && !isStringIp(pcs[0])) {
+        return 'Invalid address. x.x.x.x or localhost accepted';
+    }
+    const port = parseInt(pcs[1], 10);
+    if (!/^\d+$/.test(pcs[1])) {
+        return 'Invalid port, must be positive integer';
+    }
+    if (port <= 0 || port > 65535) {
+        return 'Invalid port, must be in range: 1-65535';
+    }
+    return '';
+};
+
+export type inputValidator = (value: string) => string | null;
